Mute Video fixing other bugss

+++ src/Containers/Socket/hostListeners.js
@@ -17,25 +17,29 @@ import socketIo from './socketListeners'
 
 export const onParticipantJoinToParticipantList = (socketIo)=>{
   socketIo.on('userHasJoinMeeting', async(msg) => {
-      console.log("SOOOOKKKEETTT userHasJoinMeeting>>>>", msg)
+      console.log("SOOOOKKKEETTT ADD PARTICIPANT userHasJoinMeeting>>>>", msg)
       const {listParticipant} = store.getState().streaming
       let list = [...listParticipant]
+      let isExist=list.filter(e=> e.userId === msg.userId)
       // const isUser=list.filter(r=> r.userId == msg.userId)
-      const {listWaitingRoom,listOnJoining} = store.getState().streaming
+      if(getSession(AppConfig.sessionMeeting).role === 'host'){
+        const {listWaitingRoom,listOnJoining} = store.getState().streaming
       
-      let lst=[...listOnJoining]
-      store.dispatch(SocketActions.removeFromJoiningList({listOnJoining:lst.filter(e=>e != msg.userId)}))
-      
-      store.dispatch(SocketActions.putToJoiningList({
-        listWaitingRoom: _.flatten(listWaitingRoom).filter(e=>e.userId != msg.userId)
-      }))
+        let lst=[...listOnJoining]
+        store.dispatch(SocketActions.removeFromJoiningList({listOnJoining:lst.filter(e=>e != msg.userId)}))
+        
+        store.dispatch(SocketActions.putToJoiningList({
+          listWaitingRoom: _.flatten(listWaitingRoom).filter(e=>e.userId != msg.userId)
+        }))
+      }
 
-
-      msg.role=msg.role.toLowerCase()
-      list.push({fullName:msg.fullName,role:msg.role,userId:msg.userId})
-      store.dispatch(SocketActions.addParticipant({
-        listParticipant: _.flatten(list)
-      }))
+     if(_.isEmpty(isExist)){
+        msg.role=msg.role.toLowerCase()
+        list.push({fullName:msg.fullName,role:msg.role,userId:msg.userId})
+        store.dispatch(SocketActions.addParticipant({
+          listParticipant: _.flatten(list)
+        }))
+     }
   })
 }
